libidset: refactor to prepare for expansion
Make the following changes to libidset in preparation
for adding more code:

- namespace the include guard symbol with FLUX_
- create idset_private.h
- add common validate_idset_flags() function
- make idset_set() a public function
- move idset_decode() to idset_decode.c
- move idset_encode() to idset_encode.c
- drop internal magic for user-after-free
- rename ENCODE_CHUNK to IDSET_ENCODE_CHUNK
- use default size on idset_create (size=0)

Note on idset_create():

idset_create (0, IDSET_FLAG_AUTOGROW) is similar to
nodeset_create(), in that it creates an idset with a
default internal size of 1K that is grown to fit id's
>= size as they are set with idset_set ().

idset_create (N, IDSET_FLAG_AUTOGROW), where N > 0, is
similar to nodeset_create_size (N).

idset_create (0, IDSET_FLAG_AUTOGROW), followed by
idset_set () is the equivalent to nodeset_create_rank ().
Similarly, nodeset_create_range () is replaced with
idset_create () followed by idset_set_range ().
(idset_set_range() will be added in a future commit).

Note on idset_encode():

idset_encode (idset, IDSET_FLAG_BRACKETS, IDSET_FLAG_RANGE)
is similar to nodeset_string (ns), except that the returned
string must be freed by the caller, and the flags to the
encode function, rather than settings on the nodeset,
determine whether the idset is encoded with brackets or ranges.
