+++ src/common/libidset/Makefile.am
@@ -13,7 +13,10 @@ AM_CPPFLAGS = \
 
 noinst_LTLIBRARIES = libidset.la
 fluxinclude_HEADERS = idset.h
-libidset_la_SOURCES = idset.c
+libidset_la_SOURCES = idset.c \
+		      idset_private.h \
+		      idset_decode.c \
+		      idset_encode.c
 
 libidset_la_CPPFLAGS = \
 	$(AM_CPPFLAGS)
