+++ src/common/libidset/idset.h
@@ -22,13 +22,14 @@
  *  See also:  http://www.gnu.org/licenses/
 \*****************************************************************************/
 
-#ifndef _IDSET_H
-#define _IDSET_H
+/* An idset is an internally sorted set of non-negative integers  (0,1,2,3...).
+ */
+
+
+#ifndef FLUX_IDSET_H
+#define FLUX_IDSET_H
 
 #include <sys/types.h>
-
-/* An idset is a numerically sorted set of non-negative integers  (0,1,2,3...).
- */
 
 enum idset_flags {
     IDSET_FLAG_AUTOGROW = 1, // allow idset size to automatically grow
@@ -37,8 +38,9 @@ enum idset_flags {
 };
 
 /* Create/destroy an idset.
- * Set the initial size to 'size' (max id is size - 1)
- * 'flags' may include IDSET_FLAG_AUTOGROW.
+ * Set the initial size to 'size' (0 means implementation uses a default size).
+ * If 'flags' includes IDSET_FLAG_AUTOGROW, the idset is resized to fit if
+ * an id >= size is set.
  * Returns idset on success, or NULL on failure with errno set.
  */
 struct idset *idset_create (size_t size, int flags);
@@ -55,7 +57,12 @@ char *idset_encode (const struct idset *idset, int flags);
  */
 struct idset *idset_decode (const char *s);
 
-#endif /* !_IDSET_H */
+/* Make id a member of idset.
+ * Return 0 on success, -1 on failure with errno set.
+ */
+int idset_set (struct idset *idset, unsigned int id);
+
+#endif /* !FLUX_IDSET_H */
 
 /*
  * vi:tabstop=4 shiftwidth=4 expandtab
